pyproject.toml
[tool.poetry]
name = "hotels"
version = "0.1.0"
description = ""
authors = ["ghw94 <ghw94@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.15.1"
black = "^24.3.0"
isort = "^5.13.2"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
pytest = "^8.1.1"
pytest-django = "^4.8.0"
mypy = "^1.9.0"
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"
flake8 = "^7.0.0"
pre-commit = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "hotels_service.settings"
